Update test suites
