+++ package.json
@@ -10,37 +10,34 @@
   },
   "scripts": {
     "build": "shx rm -rf dist && git rev-parse HEAD > .commithash && rollup -c && shx cp src/rollup/types.d.ts dist/rollup.d.ts && shx chmod a+x dist/bin/rollup",
-    "build:test": "shx rm -rf dist && rollup -c --configTest && shx cp src/rollup/types.d.ts dist/rollup.d.ts && shx chmod a+x dist/bin/rollup",
+    "build:cjs": "shx rm -rf dist && rollup -c --configTest && shx cp src/rollup/types.d.ts dist/rollup.d.ts && shx chmod a+x dist/bin/rollup",
     "build:bootstrap": "dist/bin/rollup -c && shx cp src/rollup/types.d.ts dist/rollup.d.ts && shx chmod a+x dist/bin/rollup",
     "ci:lint": "npm run lint:nofix",
-    "ci:test": "npm run build:test && npm run build:bootstrap && npm run test:all",
-    "ci:test:only": "npm run build:test && npm run build:bootstrap && npm run test:only",
-    "ci:coverage": "npm run build:test && nyc --reporter lcovonly mocha && codecov",
+    "ci:test": "npm run build:cjs && npm run build:bootstrap && npm run test:all",
+    "ci:test:only": "npm run build:cjs && npm run build:bootstrap && npm run test:only",
+    "ci:coverage": "npm run build:cjs && nyc --reporter lcovonly mocha && codecov",
     "lint": "npm run lint:ts -- --fix && npm run lint:js -- --fix && npm run lint:markdown",
     "lint:nofix": "npm run lint:ts && npm run lint:js && npm run lint:markdown",
     "lint:ts": "tslint --project .",
     "lint:js": "eslint test/test.js test/*/index.js test/utils.js test/**/_config.js",
     "lint:markdown": "markdownlint --config markdownlint.json docs/**/*.md",
-    "perf": "npm run build:test && node --expose-gc scripts/perf.js",
+    "perf": "npm run build:cjs && node --expose-gc scripts/perf.js",
     "perf:debug": "node --inspect-brk scripts/perf-debug.js",
     "perf:init": "node scripts/perf-init.js",
     "prepare": "npm run build",
     "prepublishOnly": "npm ci && npm run lint:nofix && npm run security && npm run build:bootstrap && npm run test:all",
-    "pretest": "npm run build:test",
-    "pretest:browser": "npm run build",
-    "pretest:coverage": "npm run build:test && shx rm -rf coverage/*",
-    "pretest:typescript": "shx rm -rf test/typescript/dist && shx cp -r dist test/typescript/",
-    "security": "# npm audit # deactivated until there is a solution for the lodash issue",
-    "test": "npm run test:all",
-    "test:all": "npm run test:only && npm run test:typescript && npm run test:leak && npm run test:package",
-    "test:coverage": "nyc --reporter html mocha test/test.js",
+    "security": "npm audit",
+    "test": "npm run build && npm run test:all",
+    "test:cjs": "npm run build:cjs && npm run test:only",
+    "test:quick": "mocha -b test/test.js",
+    "test:all": "npm run test:only && npm run test:browser && npm run test:typescript && npm run test:leak && npm run test:package",
+    "test:coverage": "npm run build:cjs && shx rm -rf coverage/* && nyc --reporter html mocha test/test.js",
+    "test:coverage:browser": "npm run build && shx rm -rf coverage/* && nyc mocha test/browser/index.js",
     "test:leak": "node --expose-gc test/leak/index.js",
     "test:package": "node scripts/test-package.js",
     "test:only": "mocha test/test.js",
-    "test:quick": "mocha -b test/test.js",
-    "test:typescript": "tsc --noEmit -p test/typescript && tsc --noEmit",
-    "test:browser": "mocha test/browser/index.mjs",
-    "test:browser:quick": "mocha -b test/browser/index.mjs",
+    "test:typescript": "shx rm -rf test/typescript/dist && shx cp -r dist test/typescript/ && tsc --noEmit -p test/typescript && tsc --noEmit",
+    "test:browser": "mocha test/browser/index.js",
     "watch": "rollup -cw"
   },
   "repository": "rollup/rollup",
