+++ test/browser/samples/renormalizes-external-paths/_expected/entry.js
@@ -1,0 +1,2 @@
+import './ext.js';
+import './nested/nested-ext.js';
