+++ test/browser/samples/supports-hashes/_config.js
@@ -3,10 +3,14 @@ const { loader } = require('../../../utils.js');
 module.exports = {
 	description: 'bundles files for the browser',
 	options: {
+		input: ['main', 'dep'],
 		plugins: loader({
 			main: `import {foo} from 'dep';
 console.log(foo);`,
 			dep: `export const foo = 42;`
-		})
+		}),
+		output: {
+			entryFileNames: '[name]-[hash].js'
+		}
 	}
 };
