+++ test/browser/samples/supports-hashes/_expected/dep-cf8755fa.js
@@ -1,0 +1,3 @@
+const foo = 42;
+
+export { foo };
