+++ test/browser/samples/supports-hashes/_expected/main-53604712.js
@@ -1,0 +1,3 @@
+import { foo } from './dep-cf8755fa.js';
+
+console.log(foo);
